# Avoid some cgoCheckPointer calls