@jwebb 2026-02-16
I haven't rechecked the issue, but just to clarify the monster I refer to above is `cgoCheckPointer` - it's moved a bit since I raised this.